singlenton nodeloadclient
